requirements.txt
Flask
Redis
Flask-Session
python-memcached